SAM-2087 replace use of single-precision floating point numbers with doubles
git-svn-id: https://source.sakaiproject.org/svn/common/trunk@121258 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ common/import-impl/src/java/org/sakaiproject/importer/impl/importables/AssessmentQuestion.java
@@ -43,7 +43,7 @@ public class AssessmentQuestion extends AbstractImportable {
 	private Set correctAnswerIDs;
 	private String feedbackWhenCorrect;
 	private String feedbackWhenIncorrect;
-	private Float pointValue;
+	private Double pointValue;
 	private Integer position;
 
 	public Integer getPosition() {
@@ -114,11 +114,11 @@ public class AssessmentQuestion extends AbstractImportable {
 		this.choices = choices;
 	}
 
-	public Float getPointValue() {
+	public Double getPointValue() {
 		return pointValue;
 	}
 
-	public void setPointValue(Float pointValue) {
+	public void setPointValue(Double pointValue) {
 		this.pointValue = pointValue;
 	}
